+++ tests/testthat/test-cvCovEst.R
@@ -210,3 +210,44 @@ test_that("cvCovEst's outputs are of the correct dimensions", {
   expect_true(ncol(without_true_covmat$cv_df) == 4)
   expect_true(is.null(without_true_covmat$cv_oracle_riskdiff_ratio))
 })
+
+test_that("cross-validated covariance selector handles sparse, true cov mat", {
+  expect_silent(cvCovEst(
+    dat = dat,
+    estimators = c(linearShrinkEst, linearShrinkLWEst,
+                   thresholdingEst, sampleCovEst, bandingEst,
+                   taperingEst, nlShrinkLWEst, denseLinearShrinkEst,
+                   scadEst, poetEst, adaptiveLassoEst),
+    estimator_params = list(
+      linearShrinkEst = list(alpha = c(0.1, 0.9)),
+      thresholdingEst = list(gamma = c(0.2, 2)),
+      bandingEst = list(k = c(1L, 5L)),
+      taperingEst = list(k = c(2L, 6L)),
+      scadEst = list(lambda = c(0.1, 0.2)),
+      poetEst = list(lambda = c(0.1, 0.2), k = c(1L, 2L)),
+      adaptiveLassoEst = list(lambda = c(0, 0.5), n = c(0, 0.5))
+    ),
+    cv_scheme = "v_fold", mc_split = 0.5, v_folds = 5,
+    center = TRUE, scale = FALSE, parallel = FALSE,
+    true_cov_mat = as(Sigma, "dgeMatrix")
+  ))
+  expect_silent(cvCovEst(
+    dat = dat,
+    estimators = c(linearShrinkEst, linearShrinkLWEst,
+                   thresholdingEst, sampleCovEst, bandingEst,
+                   taperingEst, nlShrinkLWEst, denseLinearShrinkEst,
+                   scadEst, poetEst, adaptiveLassoEst),
+    estimator_params = list(
+      linearShrinkEst = list(alpha = c(0.1, 0.9)),
+      thresholdingEst = list(gamma = c(0.2, 2)),
+      bandingEst = list(k = c(1L, 5L)),
+      taperingEst = list(k = c(2L, 6L)),
+      scadEst = list(lambda = c(0.1, 0.2)),
+      poetEst = list(lambda = c(0.1, 0.2), k = c(1L, 2L)),
+      adaptiveLassoEst = list(lambda = c(0, 0.5), n = c(0, 0.5))
+    ),
+    cv_scheme = "v_fold", mc_split = 0.5, v_folds = 5,
+    center = TRUE, scale = FALSE, parallel = FALSE,
+    true_cov_mat = as(Sigma, "dgCMatrix")
+  ))
+})